project_2/src_code/requirements.txt
# Amazon Review Analysis Pipeline - Dependencies
# ===============================================

# Core Data Processing
pandas==2.3.3
numpy==2.3.4
python-dotenv==1.1.1
pyyaml==6.0.3
pandasql==0.7.3

# Database Connectors
pymongo==4.15.3          # MongoDB
snowflake-connector-python>=3.0.0  # Snowflake

# AWS
boto3==1.40.56
botocore==1.40.56
s3transfer==0.14.0

# Utilities
requests==2.32.5
colorama==0.4.6
